oidc: add a nice error message for a missing flag
The kubectl plugin API changed in version 1.12 to use a model
that more resembles git.  This made the old versions of oidc
helper incompatible with the new plugin mechanism.

While the incompatibility is not fixed yet, at least this will explain
to the user what they need to do if things don't work.

There isn't a standardized way in cobra (and pflag) to mark flag as
required, so did this instead.

Change-Id: I0eeb28368b63f56776d18b6494e2cfc1ab43eba7