selectors: factor out to a separate package
Having selectors in the transform package made it difficult to reuse
them in tests as inadvertently you could make an import loop through
the visitor/testing package.

This is step 1 of 2 in making the Data object comparable.

analyzer: stop ignoring Data by default

After a refactor, it is now possible not to ignore the Data field.
Current tests still do because the contents are really irrelevant for
those outcomes.

Bug: 118512275
Change-Id: Id477fbcb7ae648b902f9dd73cf726b98118a31d3