validator: forbid annotations on NamespaceSelector
Make it an error to add nomos.dev/cluster-selector annotation on a
NamespaceSelector object.  This is to ensure that all clusters can compose
policies in exactly the same way.

A few fixes are laced into this commit that I was too lazy to factor out:

- Custom String() methods for ast objects are renamed to PrettyPrint, first
because linter says they should be named that way, and second because the
default String() is used in debugging, and omitting data to make output
pretty hampers debugging.

- validationvisitor could rely more on the underlying base visitor.
Done that and removed a few redundant methods as result.

- added logging in visitorbase.go.  Debugging visitor issues is really
convoluted without it.  Yes, I could have removed those logs once done, but
I think it may be useful some other time too.

Bug: 118314722
Change-Id: I0d83ebf0cd65fd20b3460aab459df2a5f938200a