filesystem: removes repetition in test templates
Replaces repeated annotations snippets with a pipeline call.

Since there doesn't seem to be a way to insert a bunch of templates
together when starting from strings, I did this by first generating a
template from a meta-template and inserting library code.  Feels like a
hack, but luckily needs to be done only once.

Change-Id: I0aac89d3e0cdc992f99061008435715e1bf620f6