nomosvet: add clusterselector validation
Implements object and selector validation in nomosvet.

- A defined cluster that matches no selectors is flagged.
- A cluster selector that matches no cluster is flagged.
- An object or a cluster object that references a selector that is not
defined is an error.
- An object or a cluster object that references a selector that is not
defined is an error.

These errors are not errors when the parser is runninng in a cluster,
because the parser there sees only the objects that are selected to
appear on that cluster.  For this reason, a "vet" mode is introduced,
that does a stricter repo-wide check.

Bug: 115778215
Change-Id: Ia08a634a950a87e9f7d851e1840c276a68b4f0c6