parser: use options for parser configuration
Two reasons:

(1) the true, false trailing parameters on the parser constructor are not
very informative.

(2) For cluster selection I'll need a config option which will be very
different when parser is used in nomosvet, versus in the cluster.  It would
seem awkward for this option to become another nameless bool.

Bug: 115778215
Change-Id: Ic66fc1cc5a983960436e0b6f44a0964ce44c604a