syncer: track metrics and remove unused metrics
The controller-runtime library does not expose the queue size for the
controller. We would have to use something like reflection to access it,
maybe. Also, we have not used this metric for a very long time. So, I've
removed that metric.

Added metrics for errors encountered when updating the cluster.

Change-Id: Id5325f1eaf0cf8f0b3474c78ef6e47d3476c7507