List policies only at importer startup.
This reverts parts of commit 97977d512d9a2236f3f2315cc8df90489e62d615.
Specifically, it reverts the change to load policies inside the loop,
because that appears to have caused some flakiness.

This leaves in place the change to how we delete Syncs, which is more
critical and hopefully less likely to be responsible for the flakiness.

Bug: 119284380
Change-Id: I76e4e8c9d8622cb319975da7cac78c0a61f5eb14