importer: error when seeing unexpected resource
Importer caches the list of resources on the cluster it gets from
Discovery API. It's possible for users to add or remove CRDs after this
happens. This can lead to the importer trying to load resources from
git that do have definitions on the cluster, but do not exist in the
cached list.

Previously importer would panic in this case. This changes it to return
an error.

Bug: 119278434
Change-Id: I67d4978dd96a9d19a1fc595aa338ff53250e5fcb