[bespin] Small modifications to enable visitors
This change makes the small modifications required to allow nomosvet
-print to successfully print the output of a sample Bespin repository.
Surprisingly little was required, which is awesome (and a little
worrying ;) ).

In order to work, the Bespin visitor needs to be run before the input
validator visitor, which seems to make Nomos-specific assumptions. This
is being tracked in b/119825336 and will be the next thing to do.

Fixes: 119668694

Change-Id: I3b0482386de6b5f840fa9321be3c1c5909745dcb