Use relative path for info.Source fields.
On construction, a resource.Info contains the absolute path of its
file. The absolute path is unnecessary, and we never want to display an
absolute path to users. Also, it would make automated testing difficult
if we displayed the path from root. In every instance we actually use
the path, we calculate the path relative to the repository.

This CL replaces .Source field with the relative path, eliminating the
need to constantly recalculate the relative path.

Bug: 120022947

Change-Id: I13939d76d000b8325aa54dacc4e1a9db46942e61