[bespin] Add project update support to project controller
This makes bespin project controller be able to handle the update
scenario. The problem here is we don't store terraform states
on GKE for already-created projects, when a project-update Reconcile
Request comes in, we will create a new terraform config for the request
which causes a "Project Creation" by terraform because terraform
doesn't have the state to know this is actually "updating existing
GCP resources" rather than "creating new GCP resources".

The idea is to run 'terraform import' with an empty resource
block in Reconcile(), which establishes the connection between the
terraform resource and GCP project and create a local terraform state.
After that terraform is able to correctly update the project in GCP.

Miscs:
  * This CL does a overhaul of util package, hoping to make the code
    cleaner and better reusable for future adapter development.

End2end test done on GKE cluster.

b/120000652

Change-Id: I41e7d1a14e59a2d02915ffcf2b7897747c0d20c9