Let multierror.Builder handle nil reasonably.
Right now we have to gate calls to Builder.Add with an err!=nil check
because Builder.Add throws an error if it is passed nil. This
unnecessarily complicates logic when there is no side effect to an error
having occurred other than storing it for later.

This CL changes the behavior of Builder.Add to simply do nothing if it
is passed a nil error, just like errors.Wrapf and similar do when passed
nil for their error argument. This CL changes all obvious locations
where a branch was needed previously just because this method needlessly
threw an error before.

This saves 2 lines per occurrence, and an entire logical branch of
cognitive load.

Bug: 120148065

Change-Id: Ibba92913e02241fb446bcb8e58c14642eee28aec