Validate labels and annotations for all objects
We only validate that objects in namespaces/ have valid labels and
annotations. This is an oversight, as it allows users to specify illegal
labels and annotations in objects outside namespaces/.

This CL moves the validation logic into the central loop that looks at
all objects rather than the visitor that only checks objects in
namespaces/.

This is a refactoring CL and there are no behavior changes.

Bug: 119569801
Change-Id: I02e6432370ae075ce35e56698524e9d3391d0a5f