Make ast.Visitor interface a little more sensible.
This changes the return values for Visit* methods in ast.Visitor
interface.  In retrospect, it doesn't make sense to return an interface
instead of the exact type since the visitor isn't going to be returning
a different type than the one visited.  This also removes the Node
interface since we aren't actually using it for anything practical.

Change-Id: I29ba6b4ca5b10d3162c0b1dff5adde11d7127e60