Add examples of every reproducible error
Right now there are many errors we check for that we don't have a
physical example of. This means we can't ensure they are actually
triggerable by real repositories, and we haven't spot-checked the
messages to make sure they make sense. The existing tests are in
Makefile.e2e which is quickly becoming unmaintainable.

This CL adds examples of many errors that were previously missing, and
makes a few minor logic changes in cases where errors were untriggerable
or showed wrong information. For testing, this CL adds a unit test which
cycles through example error directories to ensure they trigger the
proper errors and they are formatted reasonably.

KNV1004 is not as complete as desirable, but the fix is out of scope.
b/120236432

KNV1012 is not as complete as desirable, but the fix is out of scope.
b/120237719

KNV1014 has no example as it is untriggerable without refactoring that
would be out of scope for this CL.
b/120229144

KNV1021 has no example as there is no repeatble way to retrigger it.

KNV1022 shows incorrect information, but the fix is out of scope.
b/120237720

Bug: 119618811

Change-Id: I1de05bf0b5d9b84b2a5848eb1a263d72f56200ad