[bespin] Add Folder controller
This is similar to Project. After creating a new Folder, the
GCP-generated ID is not populated back in this CL. I am working on
a separate CL to do the GCP -> API server update: fetching the resource
from GCP, and update the resources in k8s API server accordingly.

E2E tests work for Folder create and update.

b/117158137

Change-Id: Ib646a92d72a2ce9a161d2521b684aed81dd35697