Add error for invalid metadata.name
Right now we don't ensure CRDs have a valid metadata.name, so it is
possible to create configurations that will not upload via kubectl apply
but pass nomos vet. kubectl apply requires CRDs to have a metadata.name
that is a valid DNS1123 subdomain.

This CL adds a validation error for when metadata.name is not a valid
DNS1123 label for common CRDs. This does not solve the general case of
validating user-defined CRDs - that is a much harder problem.

Bug: 120447431

Change-Id: I48d9cf942237f65a22b31c4a8bccbd3d0051f4dc