Build templating for errors
Right now we don't have pages that document individual errors. This
isn't ideal as it means users have no where to go to which says what
errors mean or what to do beyond the error message.

This CL begins an effort to automatically document error messages, and
documents two of the existing 30-ish errors we have. Many of the errors
are not trivially-documentable as instantiating them requires a client
which can interact with Kubernetes, so that refactoring is out of scope.

See this hotlist for all documentation issues I'm aware of. Feel free to
create bugs for more, as documentation issues not introduced in this CL
are out of scope: https://b.corp.google.com/hotlists/1331972

Exempt-From-Owner-Approval: Got verbal approval from frankf to submit.

Bug: 120484015
Change-Id: I9740d342fffe054870b549b38fc461f5a4a7ac76