Move errors to their own package.
Errors not having a dedicated package began causing dependency cycles as
parts of the code which wanted to throw a validation error were also
used by other classes in the validation package. The errors.go file is
also becoming unweildy, and we want to split it per-error to make things
easier to code review, look at, and create.

This CL adds a package just for validation errors. It gives a single
location that shouldn't have many dependencies.

Literally this CL is just
1) Moving errors.go to vet
2) All other changes necessary to make the code work as it did
previously

This is required to refactor parser.go. Much of its length is not
fixable until the import loops aren't an issue.

Bug: 120157756
Change-Id: Ie6b3892885796e6baddae373438e5986b0b705f7