parser.go restructuring Part 1
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This CL extracts most clusterregistry/-specific processing and
validation into its own file. The clusterregistry logic in processdirs
will be extracted later.

Change-Id: Idaf300987fbc04f79be492823fad8fd721d42076