Refactoring semantic validation Part 1
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This refactoring CL removes duplicate directory detection from
parser.go, one of many semantic validation checks that can be extracted.
The check was previously being done in two locations in parser.go, a
testament to the cognitive load of understanding the file.

There are many such semantic validation checks to factor out. The above
is the only one in the scope of this CL.

Bug: 120908722
Change-Id: Iff57658276d81de802d4d2f1af32c04e76bcf193