Refactoring semantic validation Part 5
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This CL completely extracts Sync validation from parser.go. This allowed
refactoring how Sync validation is done, so all semantic validation
takes into account the global state.

This CL improves error messages for Syncs. Defining a Namespace Sync now
only returns an error saying that the definition is illegal rather than
two errors. In the case of multiple versions for the same Kind, we now
display unsupported versions as well.

Change-Id: I0ff7992d9e3f613f22b163f2a3562b7333cb0706