Refactoring semantic validation Part 4
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This is an intermediate cleanup CL that unifies many of the things in
this chain of refactoring CLs. It's main purpose is to make
ast.FileObject the sole unit of interaction. All direct references to
runtime.Object and resource.Info have been eliminated from parser.go.

This CL also does minor refactoring things like removing reliance of
methods on being member-methods of Parser when unnecessary.

Change-Id: I39b0c3f0dd2770040c8b21ec019a528391ab101c