parser.go restructuring Part 2
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This CL is mainly copy/pasting code to different files.

This CL breaks out large pieces of parser.go which are specific to
certain directories in Nomos directories. This focuses parser.go much
more on the actual parsing process rather than all of the peripherals.

Each top dir now has its own file for parsing and validation logic
specific to it. This allowed renaming SystemOnlyResourceValidator to
DisallowSystemObjectsValidator since there is a clear spot for finding
validation for each directory.

Change-Id: Ie0fb57933da0db45a18606727f7cdc9b687a6628