Refactoring syntax validation Part 2
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

In this refactoring CL, most instances where parser.go does syntax
validation on resourece.Infos has been factored into its own file.
Refactoring validation on other object types and on semantics is out of
scope of this CL.

This CL adds many lines of the form
syntax.(name)Validator.Validate(type). This will be cleaned up in a
future step after global state validation is refactored.

Bug: 120157756
Change-Id: Iac1555cd6640de2349f7a3d15b0aeff27219efa9