Refactoring semantic validation Part 3
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This CL factors out the semantic validations in the improperly-named
validateDuplicateNames method. The method now contains no loops and is
much more readable.

There are many such semantic validation checks to factor out. The above
are the only in the scope of this CL.

Change-Id: Ia3bf4242572b7c8ea1a84c92abae79f32da7d79a