Refactoring semantic validation Part 2
The file parser.go has become increasingly unweildy as we add validation
logic to it. It is becoming harder to see how different validation steps
occur and find the ones that do exist. It needs to be broken up.

This CL factors out three semantic validation checks done in
processSystemDir.

There are many such semantic validation checks to factor out. The above
are the only in the scope of this CL.

Bug: 120908722
Change-Id: Iad2d1780a0b046eb5396abbb1511cbca57de9c1c