[bespin] Check panics for missing attachment points
This change fixes the panic check for a missing attachment point. The
attachment point actually could never be nil. It's a struct inside a
pointer, and the pointer is never nil at this point in the call
hierarchy (nil would be thrown out earlier by Nomos), but the struct inside can be
empty. So it checks for the Kind instead.

Fixes b/119616238

Change-Id: Ifc6904752bd44d1d4ef4b6eadedd33ecfe26993e