Remove splitting on if(root) for logic
While creating the directory tree in memory we've had to split out logic
specifically pertaining to creating the root. Since we know that the
first directory added is guaranteed to be the tree, we can just run the
code for creating the root if it does not yet exist. There isn't user
input that can cause multiple roots to be inserted in the tree.

This CL removes this logic that splits on whether the tree's root is
being created. It greatly simplifies the tree creation logic, removing
three instances where we would formerly panic().

Bug: 119268268

Change-Id: Id165e2bd9534df65a6aad12d062633411f60880d