Restructure nomos CLI
It isn't good practice to have the nomos main.go in the root nomos/
directory.

This CL restructures the Cobra CLI code so that main.go is in
nomos/cmd/nomos, and its dependencies are separated into their own
packages.

Most of this CL is copy/paste. There are no CLI output changes.

Bug: 120155478

Change-Id: Ib6f26506acf5d6d017166d124c89cda584761fee