Create KNV Error testing library
Right now we don't have an easy way in tests to ensure code is throwing
the correct errors, or is throwing unnecessary errors. Since we aren't
validating exactly which errors are thrown by configuration setups,
there could be unexpected or confusing errors being thrown.

This CL extracts error code validation from parser.go into its own
library. This results in many parser.go changes as the expected errors
for dozens of tests increased - we are no longer looking for the
existence of an expected error, but are validating all of them.

This CL DOES  NOT modify the tests to ensure they *only* test what
they're trying to test. In dozens of cases the tests are unintionally
testing multiple things at once, and should be corrected in a later
cleanup CL.

Bug: 121371677
Change-Id: I4c068a595cfbd2855a81cd2363a446a447135b34