Define Validator interfaces
We don't have a standard interface that we can use for validators, which
means adding code clogs up parser.go and surrounding areas.

This CL defines the reusable Validator interface which is for any object
wich adds errors for invalid Nomos repo configurations.

As a demonstration, this CL refactors the validators in sync/ to conform
to the Validator interface. Other validators will be factored to use
this interface in later CLs.

One result of this refactoring is that all Sync validations are now unit
testable, and no longer require literally creating files in order to
test the code. Thus, much of this CL is test code.

Bug: 121411328

Change-Id: I1868ade8764d1e8a87e3d3527c38287f35910b3f