Add common types to pkg/kinds/kinds.go
Right now we constantly reuse literal strings to reference Kinds. This
is unsafe as it requires that we constantly remember exactly which Group
and API version to reference when creting a Kind. It isn't obvious
without digging through other sections of code which Group and Version
to use, as some examples are inconsistent or wrong.

This CL adds Kinds to the already-started types.go, ensuring we always
are referencing the same thing when referring to a Kind.

Change-Id: I1ca7f9a696edec6720610f5d702bd82f245da0d0