Operator Make targets update for new operator rules
Previously, the operator manager manifest pointed to the lastest tag.
https://gke-internal-review.googlesource.com/c/cluster-lifecycle/cluster-operators/+/34640
changes this so that the manager yaml will have the IMAGE_TAG as a
variable, so we need to populate it appropriately.

Also, there is now a well-defined notion of "latest" for the operator,
which is the manifest at gcr.io/nomos-release/operator-latest
This change updates the make deploy-operator rule to use this operator.

Also, a small update to the example yaml to reflect the state of the
foocorp-example branches.

Also, remove a comment that is no longer relevant as release-dev is now
a different target and part of a bot.

Change-Id: I8a949bdfcd81bd4a75342ff068257d6651bc0e3f