[bespin] Remove DO NOT EDIT from Bespin CRDs
This comment originally existed because we (I) thought that we could
maintain using Kubebuilder at the same time as working in Nomos and
CNRM. This has proven to be hard, and doesn't appear to provide the
value we (I) thought it would. I don't think we are actually doing what
the comment says we are doing anymore and we are hand-editing, so let's
just remove this.

Change-Id: I35ac3cdd78b461dcfd09e84c66d2c363511f2a6c