Improve error message for unknown Sync Resources
The current message for unknown Resource definitions in syncs is
confusing to users (see bug). It doesn't warn when the user has gotten
the version wrong, or mention the (probably common) case where they
forgot which version to use.

This CL improves the error message for when the Resource in a Sync is
unknown to the cluster by letting the user know to check spelling.

This CL also adds a new error for the case where a Kind with the same
Group/Kind is defined on the cluster, but with a different version (or
versions). We may also want to consider handling the case where just
Kind is correct, and suggesting Group/Version combinations for that.

Bug: 120242539
Change-Id: I75bf2cddafd2b6461622bfddae92108c222cca03