[bespin] Create a namespace for each project
Nomos importer creates 2 kinds of policy nodes: AbstractName and
Namespace. To create a namespce policy node it requires users to
explicitly put a "Kind: Namespace" resource definition under the
directory, which is not desirable for bespin's use case because
we only want the users to care about their GCP resources.

To automatically create a namespace for a "Kind: Project" resource,
this CL plays a trick in GCPHierarchyVisitor to update the TreeNode
type to "ast.Namespace" if there are no cluster objects defined. Later
on Nomos's syncer will automatically create a namespace for this project
with the namespace name being the policynode's name.

Tests:
  * update existing unit tests inputs to have "AbstractName" type for
    a Project tree node (which matches the current implementation) and
    verify the Project tree node is of "Namespace" type after
    GCPHierarchyVisitor.
  * e2e test: a 'project' namespace is correctly created for the project
    vs. without the fix there is no namespace created. Example:
    NAMESPACE   NAME                                             AGE
    project     project.bespin.dev/zlu-project                   5s

Bug: 121392841
Change-Id: I0b521d5344957571a8dbb3b6a12e33f9cfd92daf