Create package for metadata validators
Metadata validation applies to all objects and is a coherent, single
purpose. These validations are split into a few packages and do not yet
follow the new Validator interface.

This CL moves metadata validation logic into its own package with the
minimum changes necessary. Modifying to use the Validator interface will
be more extensive (as well as adding unit tests), and that will be the
next CL.

Bug: 122251576
Change-Id: Ia59d8c14f6c11baec2ad51e781e444b255afe49f