Create ResourceID and SyncID interfaces for Errors
Right now we require users of errors to pass a specific class containing
the information the error requires in order to instantiate the error.
This is cumbersome, and makes the call sites which use Errors verbose as
they have to call the struct which is only used in the context of
Errors.

This CL adds the ResourceID interface which requires callers to pass an
object specifying the object's name, source file, and GVK.

This CL adds the SyncID interface which requires callers to pass an
object which specifies a source file containing the invalid Sync
definition, and the GVK.

This CL makes the source field of ast.FileObject unexported to keep it
from conflicting with the Source() method on the interface. This
required defining an equality comparison. This CL moves those
definitions around so they aren't redeclared in the several locations
they are used.

To make this all work, many Error messages were modified and
standardized. Every instance where it was trivial to refactor the Error
to use the new interface has been modified.

Finally, as a cleanup this CL always capitalizes "Resource" so it is
clear we are using the K8S-specific term.

Change-Id: Ic867e2e9c0fca39e0a568087ef1ad0e3305a465c