policyimporter: pca quotas may coexist
The old code disallowed the coexistence of resource quotas that are
addressed using PCA to different clusters.  This is because vetting
check checked a simpler condition, which is that there may ever be only
a single ResourceQuota resource defined in a directory.

With PCA, while it's still true that only a single ResourceQuota should
be in a namespace, any number of resource quotas may exist in a
directory, as long as they are targeted to different clusters in the
end.

This commit adds code needed to make that happen.

Bug: 121211730
Change-Id: I8dad4722676009249156b8b29406000261c3c440