[bespin] Update CRDs
Bespin CRDs (both go and YAML definitions) are out of date. Bespin team
decided to update the CRDs to be in sycn with the definitions in CNRM
repo. The expection is with this update we won't need to update the CRDs
again before MVP.

This CL only addressed the issue on the "syntax" level, e.g. making sure
the CRDs are update to date and the code can compile. There will be
another CL to fix the code logic accordingly.

Followed CL:
 - https://team-review.git.corp.google.com/c/alphabetcloud/bespin/+/359904

Bug: 122317240

Change-Id: I3abe410f0c8311aa0484fe08437e6b5f3d2ffa89