Refactor metadata.name validation and test
Validation of metadata fields does not yet use the Validator interface,
and is untested because it relies on difficult-to-instantiate objects.

This CL changes metadata validators to use a common interface, removing
their dependence on ast.FileObject. It also adds real unit tests for
every metadata validation error which can occur.

This CL also changes duplicate name validation to detect collisions
between Resources of the same GroupKind but with different versions.

Bug: 122251576

Change-Id: I1a5956532c72754ea412a8821b5a69731c080354