Rename 'vet' package to 'veterrors'
We really just want a package whose sole purpose is handling logic
related to errors. Having the package named 'vet' is misleading. This
package is also about to explode into dozens of files, so we want it to
be aptly-named.

This CL renames the 'vet' package to 'veterrors'. It also renames
'testing' to 'veterrorstest' because having to declare the alias for
`testing` literally every file that imports it is a pain.

Change-Id: I54f96e639a1277a0d21f80a45168a88651278ce4