Split out KNV errors into their own files.
Maintaining errors.go as a single file has become laborious. When
multiple people modify it simultaneously it often causes unnecessary
merge conflict. This will only get worse with time as we want to include
documentation relating to these errors and example errors for each type.

This CL splits errors.go into 40 files, 1 for the common library code
and 39 for each of the current KNV errors. It looks big, but this is
really just a copy/paste job.

This CL lets us massively parallelize error discussions as changes to
one error won't complicate others.

Bug: 122472689

Change-Id: I2a9529b96b0a9d3bbea6fb1e7ba941345231ac16