Use register pattern for veterrrors.Error
Adding a new error currently requires modifying three different files in
order to ensure an Error is usable and documented.

This CL puts a register pattern in errors.go so that each file operates
independently. The registerer does basic validation - ensuring each
error declares a unique code and that examples of errors are of the
correct type.

The modifications to individual KNV files are boilerplate-y. They won't
be as repetitious once we actually have example errors and start
documenting them.

The repeated and basically useless comments for the error code constants
are unavoiadable for now, and would require a refactoring effort outside
the scope of this CL.

Bug: 122472689
Change-Id: Icf624a5706b499e2ec011aead38d1655b2489301