Write typesafe path library.
Reviewers constantly question whether paths are OS-specific or
slash-delimited, and whether they are relative to Nomos root.

This CL creates a library which will allow compile-time checked
verification of these claims and end these debates. This is not a
generalized library; it is specifically for paths to and within Nomos
repositories. The library is robust to changes in the current working
directory.

None of these methods are used by production code since this change was
split from the CL which consumes its methods.

Background for why this is a library:

A type by itself cannot guarantee its own consistency. Since checking
is done on instantiation, callers may assume the contract these objects
promise.

If callers can directly instantiate objects by circumventing these
methods, this contract cannot be guaranteed for those instances. So the
types must not be directly instantiable.

If callers have a way of accessing the stored path which will not throw
a compilation error if the underlying data representation changes, then
callers cannot be sure the contract will hold through time. So the
underlying data must be package-private, and access granted through
methods where the caller directly says in which form they want the data.

If a library does not allow necessary transforms on these objects
(e.g. Join), so callers resort to using literal strings to represent
paths, then this contract cannot be guaranteed on those resulting
paths. So the library must declare methods on this struct for any
needed path transformations.

Change-Id: I35f2f9c8e6e8f510532813fd537eb75b73d3f50b