Add example and explanation for KNV1002
This CL documents KNV1002 with an example error and an explanation.

This CL also removes the notion of a separate "examples" section in
favor of embedding (possibley multiple) examples in the explanation.
This allows highlighting different potential ways an error can occur and
directly show the error that would be cause. These changes required
modifying the existing error-specific code in order to maintain the same
docs.

Modifying the documentation for errors other than KNV1002 is out of
scope for this change.

Bug: 120445254

Change-Id: Id8ad1ed57dc8b50e9ba70e068a92e8a322f609d5