e2e tests for syncer apply behavior
Also fixes a bug when syncing ClusterRoles.

We are required to specify a correctly formed ClusterRole in the repo, when
importer reads it. For ClusterRoles that use label selection, an empty rules
list must be specified. However, this makes it so syncer will overwrite Rules
that get generated by the ClusterRole controller. The fix is
to set ClusterRole.Rules to nil when the Rules list is empty.

Unfortunately, this cannot be done in importer. When Syncer is
deserializing ClusterRoles in the ClusterPolicy, it will autopopulate
the Rules field to empty list, even when it is set to nil in the
ClusterPolicy. So, this special-case logic has to live inside Syncer.

Fixes: 122321330
Change-Id: I01d9811699137e49a6ada5bd2f34ab05974db62f