update syncer to apply, not create/update
This changes Syncer's behavior, making it use the same approach as
running `kubectl apply` when syncing resources. Before it would strictly
overwrite existing resources with what was in the soruce of truth. It
would not preserve changes made by other controllers that modified
fields not present in the source truth.

e2e tests were run and they passed.

Fixes: 122320032

Change-Id: I321b3457da70219a23464c25ca3043a3400ce125