Makefile: embed the clean RC version into binaries
Old code used simply expanded variables for DOCKER_RUN_ARGS
and VERSION.

This meant once DOCKER_RUN_ARGS was evaluated, it was never re-evaluated
if a rule needed to adjust the value of the variable.  This caused the
evaluated VERSION to be always an offset from a previous release instead
of being equal to NEXT_RC during the release process.  To allow a change
in VERSION to be reflected in DOCKER_RUN_ARGS, both variables needed to
become recursively expanded.

The new code will always embed an RC version (e.g. v1.2.3-rc.4).  This
could be fixed by embedding the release version instead. It seems to me
that RC version is more precise and loses no information, though perhaps
less elegant.  If we decode to the contrary, it's easy to change.

Some background on make variables:
https://ftp.gnu.org/old-gnu/Manuals/make-3.79.1/html_chapter/make_6.html

Bug: 119519120
Change-Id: I23399ec184e1b96b2fed9c39786d7ee88e1ada3a