Begin using path library
We don't ensure that when we read files that we are always passing an
OS-specific absolute path.

This CL makes all instances of reading files using a path in parser.go
make compile-time guarantees that the path is OS-specific and absolute.
It also ensures the file is explicit on whether it is using a relative
slash-separated path.

Change-Id: If55d6626d2ed4d41227774cebf285bacea73087b