Rename Source() to RelativeSlashPath()
Ensure all callers to the Source() method are aware of the form of path
they're getting. Also, ensure at compile-time that objects adhere to an
interface that they maintain a relative slash-separated path.

This is one of two parts. This is only changing the method name, but
next will be changing the return type.

Change-Id: Idd5b50e677fb42c75afe064ec6e759090c627d83