e2e: move ResourceQuota test to earlier
This may de-flake the test.  We observed that frequent cluster renames
sometimes block.  This is not expected and is a bug in its own right,
but the RQ addressing feature should work independently of that.

By pulling the test to be executed earlier, it's decoupled from prior
cluster renames.  It also then does not need to rename anything, just
needs to verify that the cluster has an expected name.

Bug: 122529010
Change-Id: Id6e6d0b1a9ec1b12b76db78b4f6d97b34acbf101