Continue typesafe path refactoring
Remove numerous unnecessary uses of literal strings for paths, opting
for usage of nomospath.Relative where possible.

Add Dir() to the Sourced interface to allow consumers access to a
nomospath.Relative of the directory containing Resources.

In general, looking for files importing "path" and "path/filepath" and
factoring that logic into nomospath.go.

Notably, all KNV errors require nomospath.Relative and either delegate
the string conversion or explicitly ask for a particular form of the
path.

Bug: 122472824

Change-Id: Ib37715ed748b1352e7cf0e5f3e477447285e292a