Add Internal error type
We don't have a veterrors.Error that represents internal errors.
Instead, we usually immediately panic(). This degrades UX, causing a
splurge of stack trace information to be printed to their screen.

In tests, panics() manifest confusingly and can be difficult to track
down. This is compounded by that the stack trace is eaten by tests
"recovering" from the errors, making debugging impossible.

This CL introduces the Internal error type and begins refactorings to
eliminate unnecessary panics from our code. Internal errors are errors
that should not happen, and if they do it's a bug.

This CL also introduces the Undocumented error type. Specifically,
these are errors that we should document but haven't yet. We should
work to eliminate these.

Note that 200ish of the changed lines are spacing changes to
annotationlinervisitor_test.go.

Bug: 122825666
Bug: 122825379

Change-Id: I2dadf0dea64b108eea97d0df9fc1e20f05148838