don't use strategic merge for custom resources
We were using the controller scheme in the applier. This scheme includes
all resources that we are syncing, including custom resources. This
changes it to use a scheme that only includes native k8s resources. This
makes it so, we'll always go down the merge-patch, not
strategic-merge-patch code path whenever we run apply on a custom
resource.

Fixes: 122740188
Change-Id: Ia505ce53bcd777d2938a10c8954054b7e767deb7