Move Namespace annotation validation to Validator
Namespace annotation validation is done hierarchically, when really it
does not rely on the policy tree structure. It is also very difficult to
directly test in this state. Similarly, by testing in the ast.TreeNode
we lose the literal Namespace resource, so we can't directly reference
it in error messages.

This CL moves Namespace annotation validation to its own validator in
metadata/, making the validation logic directly verifiable. The error
message is improved to show all information necessary to locate the
Namespace with the illegal annotation.

Bug: 120445369

Change-Id: Ib811b1e97fcc6029e59ddaf107bc96737d34550a